Add initial support for two-factor authentication.
This commit adds initial support for two-factor authentication, where
the FortiGate sends a challenge through email/SMS after initial
user/password login and expects the user to provide the challenge
as a second authentication method.

This is tested working on FortiOS 5.2, with SMS tokens. It's very likely
that not all possible token authentication setups are supported (not sure
what the "code2=" parameter does, for example).